+++ DinnerHack/Home/SecondTab/LeaveBookCVC.swift
@@ -1,0 +1,8 @@
+//
+//  LeaveBookCVC.swift
+//  DinnerHack
+//
+//  Created by saint on 2023/04/29.
+//
+
+import Foundation
